+++ src/main/java/net/rewardz/RewardzMain.java
@@ -5,7 +5,6 @@ import java.util.List;
 import java.util.Map;
 
 import net.fabricmc.api.ModInitializer;
-import net.fabricmc.loader.api.FabricLoader;
 import net.rewardz.init.CommandInit;
 import net.rewardz.init.ConfigInit;
 import net.rewardz.init.LoaderInit;
@@ -14,7 +13,6 @@ import net.rewardz.packet.RewardsServerPacket;
 
 public class RewardzMain implements ModInitializer {
 
-    public static final boolean isPatchouliButtonLoaded = FabricLoader.getInstance().isModLoaded("patchoulibutton");
     public static final Map<Integer, Map<Integer, List<Object>>> REWARD_MAP = new HashMap<Integer, Map<Integer, List<Object>>>();
 
     @Override