+++ src/main/java/net/rewardz/mixin/client/InventoryScreenMixin.java
@@ -32,15 +32,16 @@ public abstract class InventoryScreenMixin extends AbstractInventoryScreen<Playe
     @Inject(method = "mouseClicked", at = @At("HEAD"))
     private void mouseClickedMixin(double mouseX, double mouseY, int button, CallbackInfoReturnable<Boolean> info) {
         if (this.client != null && this.focusedSlot == null
-                && this.isPointWithinBounds(ConfigInit.CONFIG.posX + (RewardzMain.isPatchouliButtonLoaded ? 23 : 0), ConfigInit.CONFIG.posY, 20, 18, (double) mouseX, (double) mouseY)) {
+                && this.isPointWithinBounds(ConfigInit.CONFIG.posX + (RenderInit.isPatchouliButtonLoaded ? 23 : 0), ConfigInit.CONFIG.posY, 20, 18, (double) mouseX, (double) mouseY)) {
             RewardsClientPacket.writeC2SRewardsScreenPacket(client);
         }
     }
 
     @Inject(method = "drawBackground", at = @At("TAIL"))
     protected void drawBackgroundMixin(DrawContext context, float delta, int mouseX, int mouseY, CallbackInfo info) {
-        int rewardButtonX = +ConfigInit.CONFIG.posX + (RewardzMain.isPatchouliButtonLoaded ? 23 : 0);
-        if (this.isPointWithinBounds(ConfigInit.CONFIG.posX, ConfigInit.CONFIG.posY, 20, 18, (double) mouseX, (double) mouseY)) {
+        int rewardButtonX = ConfigInit.CONFIG.posX + (RenderInit.isPatchouliButtonLoaded ? 23 : 0);
+
+        if (this.isPointWithinBounds(rewardButtonX, ConfigInit.CONFIG.posY, 20, 18, (double) mouseX, (double) mouseY)) {
             context.drawTexture(RenderInit.REWARD_ICONS, this.x + rewardButtonX, this.y + ConfigInit.CONFIG.posY, 196, 42, 20, 18);
             renderCheckMark(context);
             context.drawTooltip(this.textRenderer, Text.translatable("screen.rewardz"), mouseX, mouseY);
@@ -53,9 +54,9 @@ public abstract class InventoryScreenMixin extends AbstractInventoryScreen<Playe
     private void renderCheckMark(DrawContext context) {
         if (this.client.player != null && ((RewardPlayerAccess) this.client.player).getUsedRewardDays().size() < ((RewardPlayerAccess) this.client.player).getRewardDayCount()
                 && !hasOnlyExactRewardsLeft()) {
-            context.drawTexture(RenderInit.REWARD_ICONS, this.x + 2 + ConfigInit.CONFIG.posX + (RewardzMain.isPatchouliButtonLoaded ? 23 : 0), this.y + 2 + ConfigInit.CONFIG.posY, 192, 0, 16, 14);
+            context.drawTexture(RenderInit.REWARD_ICONS, this.x + 2 + ConfigInit.CONFIG.posX + (RenderInit.isPatchouliButtonLoaded ? 23 : 0), this.y + 2 + ConfigInit.CONFIG.posY, 192, 0, 16, 14);
         } else {
-            context.drawTexture(RenderInit.REWARD_ICONS, this.x + 2 + ConfigInit.CONFIG.posX + (RewardzMain.isPatchouliButtonLoaded ? 23 : 0), this.y + 2 + ConfigInit.CONFIG.posY, 176, 0, 16, 14);
+            context.drawTexture(RenderInit.REWARD_ICONS, this.x + 2 + ConfigInit.CONFIG.posX + (RenderInit.isPatchouliButtonLoaded ? 23 : 0), this.y + 2 + ConfigInit.CONFIG.posY, 176, 0, 16, 14);
         }
     }