+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 - Added reward commands
 ### Fixed:
 - Fixed reward inventory
+- Fixed button position
 ### Changed:
 -